Feat!: Only use requests to download data from the connectivity tool
