+++ Code/connectivity_tool/__init__.py
@@ -1,0 +1,1 @@
+from .core import find_connected_flares
